Solved issue 121: regex error in pem.js
If there is a match, str.match(regex) always returns at least one group, where
the first group is always the entire match, and the rest are the capturing groups.
Therefore getFirstCapturingGroup returns an array of two groups, and the second
one is the content captured.

Simplified CERT_REGEX and PRIVATE_KEY_REGEX and made them stricter, also added
some comments on top of pem.js.