Issue #71
Some properties (clientID, callbackURL, responseType, responseMode) from the provided options are checked in the `authenticate` time, this is not good because:
(1) code should crash immediately instead of the time when `authenticate` is called, immediate crash gives the user immediate feedback of their configuration error
(2) these properties are checked for every request that needs `authenticate`. This is not necessary since the value of these properties would be changed.
Considering all these, we now check them at the creation time of strategy.