(Add-fix) multiple nonce and state support
We now have a unified handler for nonce and state. We save {state: x, nonce: x, policy: x, timeStamp: x} tuple into session, and we verify the state and nonce pair now instead of verifying them separately. Since we are using state and nonce together as a tuple, the maxAmount and maxAge will be the same for them.